Reorganized configuration classes
