+++ src/Nox.Core/Configuration/Api/ApiRouteConfiguration.cs
