+++ src/Nox.Core/Configuration/Api/ApiRouteResponseConfiguration.cs
