+++ src/Nox.Core/Configuration/Entity/EntityAttributeConfiguration.cs
