+++ src/Nox.Core/Configuration/Entity/EntityConfiguration.cs
