+++ src/Nox.Core/Configuration/Loader/LoaderConfiguration.cs
