+++ src/Nox.Core/Configuration/Loader/LoaderLoadStrategyConfiguration.cs
