+++ src/Nox.Core/Configuration/Loader/LoaderSourceConfiguration.cs
