+++ src/Nox.Core/Configuration/Messaging/MessagingProviderConfiguration.cs
