+++ src/Nox.Core/Configuration/VersionControl/ProjectConfiguration.cs
