+++ src/Nox.Core/Configuration/VersionControl/TeamConfiguration.cs
