+++ src/Nox.Core/Configuration/VersionControl/TeamMemberConfiguration.cs
