+++ src/Nox.Core/Configuration/VersionControl/VersionControlConfiguration.cs
