+++ src/Nox.Core/Extensions/ConfigurationExtensions.cs
@@ -1,9 +1,11 @@
 using FluentValidation;
 using Microsoft.Extensions.DependencyInjection;
+using Nox.Core.Configuration;
+using Nox.Core.Helpers;
 using Nox.Core.Interfaces.Configuration;
 using Nox.Core.Validation.Configuration;
 
-namespace Nox.Core.Configuration;
+namespace Nox.Core.Extensions;
 
 public static class ConfigurationExtensions
 {
