+++ src/Nox.Lib/Extensions/ServiceExtensions.cs
@@ -4,6 +4,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Nox.Api;
 using Nox.Lib;
 using Nox.Core.Configuration;
+using Nox.Core.Extensions;
+using Nox.Core.Helpers;
 using Nox.Core.Interfaces;
 using Nox.Data;
 using Nox.Etl;
